+++ src/components/features/protocol/credentials/view/credential-view.tsx
@@ -2,7 +2,7 @@ import useTranslation from 'next-translate/useTranslation';
 import { useRouter } from 'next/router';
 
 import ExternalLink from '@/components/atoms/external-link';
-import { MintDialogProtocol } from '@/components/organisms/mint-modal/mint-dialog-protocol';
+import { MintDialogProtocol } from '@/components/organisms/mint/mint-modal/mint-dialog-protocol';
 import ModalShareCredential from '@/components/molecules/modal/modal-share-credential';
 import { ROUTES } from '@/constants/routes';
 import { useMintData } from '@/hooks/use-mint-data';
